+++ Packages/com.thenathannator.plasticband/PlasticBand/Devices/FiveFretGuitar/GuitarHeroGuitar/XInputGuitarHeroGuitar.cs
@@ -9,15 +9,15 @@ using UnityEngine.InputSystem.Utilities;
 using UnityEngine.InputSystem.XInput;
 
 // PlasticBand reference doc:
-// https://github.com/TheNathannator/PlasticBand/blob/main/Docs/Instruments/5-Fret%20Guitar/Rock%20Band/Xbox%20360.md
+// https://github.com/TheNathannator/PlasticBand/blob/main/Docs/Instruments/5-Fret%20Guitar/Guitar%20Hero/Xbox%20360.md
 
 namespace PlasticBand.Devices.LowLevel
 {
     /// <summary>
-    /// The state format for XInput Guitar devices.
+    /// The state format for XInput Guitar Hero guitars.
     /// </summary>
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
-    internal unsafe struct XInputGuitarState : IInputStateTypeInfo
+    internal unsafe struct XInputGuitarHeroGuitarState : IInputStateTypeInfo
     {
         public FourCC format => XInputGamepad.Format;
 
@@ -34,29 +34,39 @@ namespace PlasticBand.Devices.LowLevel
         [InputControl(name = "selectButton", layout = "Button", bit = 5, displayName = "Back")]
 
         [InputControl(name = "orangeFret", layout = "Button", bit = 8)]
+        [InputControl(name = "spPedal", layout = "Button", bit = 9)]
 
         [InputControl(name = "greenFret", layout = "Button", bit = 12)]
         [InputControl(name = "redFret", layout = "Button", bit = 13)]
         [InputControl(name = "blueFret", layout = "Button", bit = 14)]
         [InputControl(name = "yellowFret", layout = "Button", bit = 15)]
-
-        [InputControl(name = "soloGreen", layout = "MaskButton", format = "USHT", bit = 0, parameters = "mask=0x1040")]
-        [InputControl(name = "soloRed", layout = "MaskButton", format = "USHT", bit = 0, parameters = "mask=0x2040")]
-        [InputControl(name = "soloYellow", layout = "MaskButton", format = "USHT", bit = 0, parameters = "mask=0x8040")]
-        [InputControl(name = "soloBlue", layout = "MaskButton", format = "USHT", bit = 0, parameters = "mask=0x4040")]
-        [InputControl(name = "soloOrange", layout = "MaskButton", format = "USHT", bit = 0, parameters = "mask=0x0140")]
         public ushort buttons;
 
-        // TODO: Define specific ranges for each of the notches
-        [InputControl(name = "pickupSwitch", layout = "Axis", parameters = "scale=true,scaleFactor=4")]
-        public byte pickupSwitch;
+        // Was gonna use these parameters based on my Les Paul, but maybe it would be best to
+        // leave that up to calibration systems rather than assuming all guitars will be the same
+        // parameters = "normalize=true,normalizeMin=0.1,normalizeMax=0.6,normalizeZero=0.36,clamp=2,clampMin=-1,clampMax=1"
+        [InputControl(name = "accelY", layout = "Axis", noisy = true, defaultState = 0x80, parameters = "normalize=true,normalizeMin=0,normalizeMax=1,normalizeZero=0.5")]
+        public byte accelY;
 
-        public fixed byte unused[5];
+        // parameters = "normalize=true,normalizeMin=0.2,normalizeMax=0.75,normalizeZero=0.45,clamp=2,clampMin=-1,clampMax=1"
+        [InputControl(name = "accelZ", layout = "Axis", noisy = true, defaultState = 0x80, parameters = "normalize=true,normalizeMin=0,normalizeMax=1,normalizeZero=0.5")]
+        public byte accelZ;
+
+        [InputControl(name = "touchGreen", layout = "GuitarHeroSlider", format = "SHRT")]
+        [InputControl(name = "touchRed", layout = "GuitarHeroSlider", format = "SHRT")]
+        [InputControl(name = "touchYellow", layout = "GuitarHeroSlider", format = "SHRT")]
+        [InputControl(name = "touchBlue", layout = "GuitarHeroSlider", format = "SHRT")]
+        [InputControl(name = "touchOrange", layout = "GuitarHeroSlider", format = "SHRT")]
+        public short slider;
+
+        public short unused;
 
         [InputControl(name = "whammy", layout = "Axis", defaultState = short.MinValue, parameters = "normalize=true,normalizeMin=-1,normalizeMax=1,normalizeZero=-1")]
         public short whammy;
 
         [InputControl(name = "tilt", layout = "Axis", noisy = true)]
+        // parameters = "normalize=true,normalizeMin=-0.85,normalizeMax=1,normalizeZero=0,clamp=2,clampMin=-1,clampMax=1"
+        [InputControl(name = "accelX", layout = "Axis", noisy = true)]
         public short tilt;
     }
 }
@@ -64,32 +74,32 @@ namespace PlasticBand.Devices.LowLevel
 namespace PlasticBand.Devices
 {
     /// <summary>
-    /// An XInput Guitar.
+    /// An XInput Guitar Hero guitar.
     /// </summary>
-    [InputControlLayout(stateType = typeof(XInputGuitarState), displayName = "XInput Rock Band Guitar")]
-    public class XInputGuitar : RockBandGuitar
+    [InputControlLayout(stateType = typeof(XInputGuitarHeroGuitarState), displayName = "XInput Guitar Hero Guitar")]
+    public class XInputGuitarHeroGuitar : GuitarHeroGuitar
     {
         /// <summary>
-        /// The current <see cref="XInputGuitar"/>.
+        /// The current <see cref="XInputGuitarHeroGuitar"/>.
         /// </summary>
-        public static new XInputGuitar current { get; private set; }
+        public static new XInputGuitarHeroGuitar current { get; private set; }
 
         /// <summary>
-        /// A collection of all <see cref="XInputGuitar"/>s currently connected to the system.
+        /// A collection of all <see cref="XInputGuitarHeroGuitar"/>s currently connected to the system.
         /// </summary>
-        public new static IReadOnlyList<XInputGuitar> all => s_AllDevices;
-        private static readonly List<XInputGuitar> s_AllDevices = new List<XInputGuitar>();
+        public new static IReadOnlyList<XInputGuitarHeroGuitar> all => s_AllDevices;
+        private static readonly List<XInputGuitarHeroGuitar> s_AllDevices = new List<XInputGuitarHeroGuitar>();
 
         /// <summary>
-        /// Registers <see cref="RockBandGuitar"/> to the input system.
+        /// Registers <see cref="XInputGuitarHeroGuitar"/> to the input system.
         /// </summary>
         internal new static void Initialize()
         {
-            XInputDeviceUtils.Register<XInputGuitar>(XInputController.DeviceSubType.Guitar);
+            XInputDeviceUtils.Register<XInputGuitarHeroGuitar>(XInputController.DeviceSubType.GuitarAlternate);
         }
 
         /// <summary>
-        /// Sets this device as the current <see cref="XInputGuitar"/>.
+        /// Sets this device as the current <see cref="XInputGuitarHeroGuitar"/>.
         /// </summary>
         public override void MakeCurrent()
         {